fix: move lib to src and dist to lib for node backwards-compatibility
