+++ examples/node/Dockerfile
@@ -4,9 +4,9 @@ WORKDIR /app
 RUN git clone https://github.com/twilio/twilio-node.git
 WORKDIR /app/twilio-node/
 
-RUN rm -rf lib/rest/ spec/
+RUN rm -rf src/rest/ spec/
 
-COPY lib lib
+COPY src src
 COPY spec spec
 
 # pipefail prevents errors in a pipeline from being masked.
