+++ examples/node/src/rest/ApiBase.ts
