+++ examples/node/src/rest/FlexApiBase.ts
