+++ examples/node/src/rest/VersionlessBase.ts
