+++ examples/node/src/rest/api/v2010/account/call/feedbackCallSummary.ts
@@ -17,7 +17,7 @@ import V2010 from "../../../V2010";
 const deserialize = require("../../../../../base/deserialize");
 const serialize = require("../../../../../base/serialize");
 import { isValidPathParam } from "../../../../../base/utility";
-import { PhoneNumberCapabilities } from "../../../../../../lib/interfaces";
+import { PhoneNumberCapabilities } from "../../../../../../src/interfaces";
 
 export class TestResponseObjectTestArrayOfObjects {
   "count"?: number;
