+++ examples/node/src/rest/flexApi/v1/credential/aws.ts
