+++ examples/node/src/rest/versionless/DeployedDevices.ts
