+++ examples/node/src/rest/versionless/deployed_devices/fleet.ts
