+++ examples/node/src/rest/versionless/understand/assistant.ts
