+++ src/main/resources/config/node.json
@@ -1,10 +1,10 @@
 {
     "library": {
         "phone_number_capabilities": {
-            "PhoneNumberCapabilities": "lib/interfaces"
+            "PhoneNumberCapabilities": "src/interfaces"
         },
         "twiml": {
-            "TwiML": "lib/twiml/TwiML"
+            "TwiML": "src/twiml/TwiML"
         }
     },
